Move memory tracking helpers
This allows the memory tracking helpers to be used in unit tests.
